ref(test): Colocate tooltip store with AcceptanceTestTooltip
